modal refactoring child content에 따라 나눔 - base modal을 필드로 가지고 있는 input, textarea modal 생성 - topic title 수정 시 데이터 초기화와 리스너 등록이 안되던 버그 해결
