+++ client/src/components/modal/base-modal/index.ts
@@ -1,0 +1,5 @@
+import './modal.scss';
+import Modal, { ModalInterface } from './modal';
+
+export default Modal;
+export { ModalInterface };
