+++ client/src/components/modal/base-modal/modal.scss
